test(workflow): relocate tools/compensation/state-store/checkpoint/migration/schemas legacy suites to co-located siblings (#1705)
